Add ImageDimensionReducer Code
Add ImageDimensionReducer that can be used to load images and do
dimensionality reduction on the images. The reduced dimensional
representation of the data can be exported to a JSON file and
thumbnails of the original images can also be generated. The file
test.py was also added as a simple example.